Inspect pmtiles
