+++ app/javascript/geo/index_map_renderer.js
@@ -1,8 +1,9 @@
 export class IndexMapRenderer {
-  constructor(map, dataAttributes, el) {
+  constructor(map, dataAttributes, openSidebarWithContent, highlightFeature) {
     this.map = map
     this.dataAttributes = dataAttributes
-    this.el = el
+    this.openSidebarWithContent = openSidebarWithContent
+    this.highlightFeature = highlightFeature
   }
 
   render(data) {
@@ -89,15 +90,6 @@ export class IndexMapRenderer {
         }
       })
     }
-
-    this.setupSidebar()
-    this.addOpacityControl(
-      [
-        { id: "index-map-fill", property: "fill-opacity" },
-        { id: "index-map-circle", property: "circle-opacity" }
-      ],
-      0.75
-    )
   }
 
   getBoundsFromGeoJSON(data) {
@@ -172,211 +164,5 @@ export class IndexMapRenderer {
     thumbPromise.then(() =>
       this.openSidebarWithContent(this.indexMapInfo(data))
     )
-  }
-
-  // Highlight a single GeoJSON feature (e.g. from an index map click).
-  // e.features[0] is a MapGeoJSONFeature (a MapLibre-specific class), not a
-  // plain object. Reconstruct it as plain GeoJSON so MapLibre can serialize it
-  // to its web worker when adding it as a GeoJSON source.
-  highlightFeature(feature) {
-    this.setHighlightData({
-      type: "FeatureCollection",
-      features: [
-        {
-          type: "Feature",
-          geometry: feature.geometry,
-          properties: { ...feature.properties }
-        }
-      ]
-    })
-  }
-
-  setHighlightData(geojsonData) {
-    const color = JSON.parse(this.dataAttributes.geoViewerColors).selected
-
-    if (this.map.getSource("highlight-source")) {
-      this.map.getSource("highlight-source").setData(geojsonData)
-    } else {
-      this.map.addSource("highlight-source", {
-        type: "geojson",
-        data: geojsonData
-      })
-
-      this.map.addLayer({
-        id: "highlight-fill",
-        type: "fill",
-        source: "highlight-source",
-        filter: ["==", ["geometry-type"], "Polygon"],
-        paint: { "fill-color": color, "fill-opacity": 0.5 }
-      })
-
-      this.map.addLayer({
-        id: "highlight-line",
-        type: "line",
-        source: "highlight-source",
-        paint: { "line-color": color, "line-width": 2 }
-      })
-
-      this.map.addLayer({
-        id: "highlight-circle",
-        type: "circle",
-        source: "highlight-source",
-        filter: ["==", ["geometry-type"], "Point"],
-        paint: {
-          "circle-radius": 8,
-          "circle-color": color,
-          "circle-opacity": 0.7
-        }
-      })
-    }
-  }
-
-  openSidebarWithContent(html) {
-    const sidebar = this.el.querySelector(".sul-embed-geo-sidebar")
-    const button = sidebar.querySelector("button")
-    const content = sidebar.querySelector(".sul-embed-geo-sidebar-content")
-    button.setAttribute("aria-expanded", "true")
-    content.classList.add("show")
-    content.innerHTML = html
-    content.style.maxHeight = `${this.el.clientHeight - 90}px`
-    content.setAttribute("aria-hidden", "false")
-  }
-
-  geoSidebar() {
-    const title = "Map Sheet"
-    const body = "Click the map to inspect a sheet."
-
-    return `<div class="sul-embed-geo-sidebar">
-                  <div class="sul-embed-geo-sidebar-header">
-                    <h3>${title}</h3>
-                    <button aria-label="expand/collapse" aria-expanded="true" aria-controls="sidebarContent">
-                      <svg class="MuiSvgIcon-root KeyboardArrowUpSharp" focusable="false" aria-hidden="true" viewBox="0 0 24 24"><path d="M7.41 15.41 12 10.83l4.59 4.58L18 14l-6-6-6 6z"></path></svg>
-                    </button>
-                  </div>
-                  <div id="sidebarContent" class="sul-embed-geo-sidebar-content show">${body}</div>
-                </div>`
-  }
-
-  setupSidebar() {
-    const sidebarControl = {
-      onAdd: () => {
-        const container = document.createElement("div")
-        container.className = "maplibregl-ctrl"
-        container.innerHTML = this.geoSidebar()
-        container.addEventListener("click", e => {
-          const button = e.target.closest("button")
-          if (!button) return
-          const contentEl = document.getElementById(
-            button.getAttribute("aria-controls")
-          )
-          const expanded = button.getAttribute("aria-expanded") === "true"
-          contentEl.classList.toggle("show", !expanded)
-          button.setAttribute("aria-expanded", String(!expanded))
-        })
-        return container
-      },
-      onRemove: () => {}
-    }
-    this.map.addControl(sidebarControl, "top-right")
-  }
-
-  // Reimplements L.Control.LayerOpacity as a vanilla-JS MapLibre IControl,
-  // reusing the existing .opacity-control CSS from geo.css.
-  addOpacityControl(layerSpecs, initialOpacity = 0.75) {
-    let mouseMoveHandler, mouseUpHandler
-
-    const opacityControl = {
-      onAdd: map => {
-        const opacityPercent = Math.round(initialOpacity * 100)
-
-        const container = document.createElement("div")
-        container.className = "opacity-control unselectable maplibregl-ctrl"
-
-        const area = document.createElement("div")
-        area.className = "opacity-area"
-
-        const handle = document.createElement("div")
-        handle.className = "opacity-handle"
-        handle.setAttribute("tabindex", "0")
-        handle.innerHTML = `
-          <div class="opacity-arrow-up"></div>
-          <div class="opacity-text">${opacityPercent}%</div>
-          <div class="opacity-arrow-down"></div>`
-
-        const bottom = document.createElement("div")
-        bottom.className = "opacity-bottom"
-
-        container.append(area, handle, bottom)
-
-        handle.style.top = `calc(100% - ${opacityPercent}% - 12px)`
-        bottom.style.top = `calc(100% - ${opacityPercent}%)`
-        bottom.style.height = `${opacityPercent}%`
-
-        // Prevent map interactions from passing through the control
-        container.addEventListener("click", e => e.stopPropagation())
-        container.addEventListener("mousedown", e => e.stopPropagation())
-
-        const updateOpacity = opacity => {
-          layerSpecs.forEach(({ id, property }) => {
-            if (map.getLayer(id)) map.setPaintProperty(id, property, opacity)
-          })
-        }
-
-        let dragStart = null
-        let dragStartTop
-
-        handle.addEventListener("mousedown", e => {
-          dragStart = e.clientY
-          dragStartTop = handle.offsetTop - 12
-        })
-
-        mouseMoveHandler = e => {
-          if (dragStart === null) return
-          const percentInverse =
-            Math.max(0, Math.min(200, dragStartTop + e.clientY - dragStart)) / 2
-          handle.style.top = `${percentInverse * 2 - 13}px`
-          handle.querySelector(".opacity-text").innerHTML =
-            `${Math.round((1 - percentInverse / 100) * 100)}%`
-          bottom.style.height = `${Math.max(0, (100 - percentInverse) * 2 - 13)}px`
-          bottom.style.top = `${Math.min(200, percentInverse * 2 + 13)}px`
-          updateOpacity(1 - percentInverse / 100)
-        }
-
-        mouseUpHandler = () => {
-          dragStart = null
-        }
-
-        document.addEventListener("mousemove", mouseMoveHandler)
-        document.addEventListener("mouseup", mouseUpHandler)
-
-        handle.addEventListener("keydown", e => {
-          const current = parseInt(
-            handle.querySelector(".opacity-text").innerHTML
-          )
-          let next =
-            e.key === "ArrowUp"
-              ? current + 1
-              : e.key === "ArrowDown"
-                ? current - 1
-                : null
-          if (next === null) return
-          e.preventDefault()
-          next = Math.max(0, Math.min(100, next))
-          handle.style.top = `calc(100% - ${next}% - 12px)`
-          handle.querySelector(".opacity-text").innerHTML = `${next}%`
-          bottom.style.height = `${next}%`
-          bottom.style.top = `calc(100% - ${next}%)`
-          updateOpacity(next / 100)
-        })
-
-        return container
-      },
-      onRemove: () => {
-        document.removeEventListener("mousemove", mouseMoveHandler)
-        document.removeEventListener("mouseup", mouseUpHandler)
-      }
-    }
-
-    this.map.addControl(opacityControl, "top-left")
   }
 }
