+++ config/settings/development.yml
@@ -114,7 +114,7 @@ sandbox:
     - label: Index map
       purl: kh860qj3406
     - label: PMTiles
-      purl: hf224mw4004
+      purl: jx110jn7648
     - label: GeoJSON points
       purl: nz577hv6134
     - label: GeoJSON lines
